requirements.txt
GoogleAppEngineCloudStorageClient==1.9.22.1
GoogleAppEngineMapReduce==1.9.22.0
GoogleAppEnginePipeline==1.9.22.1
Pillow==4.0.0
PyPDF2==1.23
PyYAML==3.11
babel==1.3
dnspython==1.11.1
ecdsa==0.13
functools32==3.2.3-2
futures==3.2.0
git+git://github.com/ojii/pymaging-png.git#egg=pymaging-png
git+git://github.com/ojii/pymaging.git#egg=pymaging
google-api-python-client==1.4.2
graphy==1.0
html2text==2019.8.11
html5lib==0.999
icalendar==3.5
lxml==3.7.3
oauth2client==1.5.2
ply==3.4
pydkim==0.3
python-dateutil==2.6.0
python-jose==3.0.1
pytz==2017.2
qrcode==5.3
reportlab==3.1.44
requests-toolbelt==0.9.1
requests==2.22.0
simplejson==3.10.0
six==1.12.0
slimit==0.7.4
wsgiref==0.1.2
xhtml2pdf==0.0.6
xlrd==1.0.0
xlwt==0.7.4
sendgrid==5.2.1
influxdb==5.0.0
createsend==5.0.0
https://github.com/our-city-app/log-offload/archive/master.zip
typing==3.6.6
pycrypto==2.6.1
Markdown==3.0.1
beautifulsoup4==4.7.1
facebook-sdk==3.1.0
stripe==2.35.1
zeep==3.4.0
autopep8
PyRSS2Gen==1.1
